+++ core/src/main/java/org/jdbi/v3/core/Handle.java
@@ -425,16 +425,13 @@ public class Handle implements Closeable, Configurable<Handle> {
     /**
      * Executes <code>callback</code> in a transaction.
      *
-     * @param callback a callback which will receive an open handle, in a transaction.
+     * @param consumer a callback which will receive an open handle, in a transaction.
      * @param <X> exception type thrown by the callback, if any
      *
      * @throws X any exception thrown by the callback
      */
-    public <X extends Exception> void useTransaction(final HandleConsumer<X> callback) throws X {
-        inTransaction(handle -> {
-            callback.useHandle(handle);
-            return null;
-        });
+    public <X extends Exception> void useTransaction(final HandleConsumer<X> consumer) throws X {
+        inTransaction(consumer.asCallback());
     }
 
     /**
@@ -480,15 +477,12 @@ public class Handle implements Closeable, Configurable<Handle> {
      * </p>
      * @param level the transaction isolation level which will be applied to the connection for the scope of this
      *              transaction, after which the original isolation level will be restored.
-     * @param callback a callback which will receive an open handle, in a transaction.
+     * @param consumer a callback which will receive an open handle, in a transaction.
      * @param <X> exception type thrown by the callback, if any
      * @throws X any exception thrown by the callback
      */
-    public <X extends Exception> void useTransaction(TransactionIsolationLevel level, HandleConsumer<X> callback) throws X {
-        inTransaction(level, handle -> {
-            callback.useHandle(handle);
-            return null;
-        });
+    public <X extends Exception> void useTransaction(TransactionIsolationLevel level, HandleConsumer<X> consumer) throws X {
+        inTransaction(level, consumer.asCallback());
     }
 
     /**