+++ sqlobject/src/main/java/org/jdbi/v3/sqlobject/Handlers.java
@@ -14,6 +14,7 @@
 package org.jdbi.v3.sqlobject;
 
 import java.lang.reflect.Method;
+import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.CopyOnWriteArrayList;
@@ -33,6 +34,7 @@ public class Handlers implements JdbiConfig<Handlers> {
 
     public Handlers() {
         register(new DefaultMethodHandlerFactory());
+        register(new WithHandleMethodHandlerFactory());
         register(new SqlMethodHandlerFactory());
         register(new BridgeMethodHandlerFactory());
     }
@@ -60,5 +62,14 @@ public class Handlers implements JdbiConfig<Handlers> {
     @Override
     public Handlers createCopy() {
         return new Handlers(this);
+    }
+
+    static Method methodLookup(Class<?> klass, String methodName, Class<?>... parameterTypes) {
+        try {
+            return klass.getMethod(methodName, parameterTypes);
+        } catch (NoSuchMethodException | SecurityException e) {
+            throw new IllegalStateException(
+                    String.format("can't find %s#%s%s", klass.getName(), methodName, Arrays.asList(parameterTypes)), e);
+        }
     }
 }